backend/requirements.txt
flask
flask_cors
flask_sqlalchemy
pandas
python-dotenv
psycopg2-binary
axios
flask-testing
celery 
redis